% $Id$

\name{DeMix-package}
\alias{DeMix-package}
\docType{package}
\title{\Sexpr[results=text]{packageDescription("DeMix")$Title}}
\description{
  \Sexpr[results=text]{packageDescription("DeMix")$Description}
}
\details{
  \tabular{ll}{
    Package: \tab \Sexpr[results=text]{packageDescription("DeMix")$Package}\cr
    Type:    \tab Package\cr
    Version: \tab \Sexpr[results=text]{packageDescription("DeMix")$Version}\cr
    Date:    \tab \Sexpr[results=text]{packageDescription("DeMix")$Date}\cr
    License: \tab \Sexpr[results=text]{packageDescription("DeMix")$License}\cr
  }

  For a complete list of functions, use \code{library(help="DeMix")}.\cr
  For a high-level summary of the changes for each revision, use\cr
    \code{file.show(system.file("NEWS", package="DeMix"))}.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\keyword{package}

