% $Id$

\name{ntot_NT_Liver-data}
\alias{ntot_NT_Liver-data}
\alias{ntot_NT_Liver}
\alias{ntot_NT_Liver.df}
\docType{data}
\title{TBD expression in cell lines}
\description{
  This data set contains the expression levels of TBD...
  The first 60 columns are normal samples, followed by 151 tumor samples.

  See corresponding manpage of the raw data for a description of the
  design of the \acronym{ntot_NT_Liver}.
}
\usage{
data(ntot_NT_Liver)
}
\details{
  The corresponding raw datafiles are available in the
  \file{extdata} subdirectory of the \pkg{DeMixSampleData} package.
}
\format{
  The object \code{ntot_NT_Liver.df} is a data.frame with 2888 rows
  by 211 columns.
}
\author{
  P. Roebuck \email{proebuck@mdanderson.org}
}
\seealso{ 
  \code{\link[DeMixSampleData]{ntot_NT_Liver-extdata}}
}
\keyword{datasets}

